refactor(egui): F0(c) — split frame.rs (835) em frame/ <500 + kill-gate de vsync
Gate read_before_commit dispara em arquivo >500 linhas. frame.rs (835) quebrado
em pasta frame/ de submódulos coesos, comportamento preservado:
- frame/gpu.rs (385) — GpuConfig/WindowChrome/Backend/SharedGpu/shared_gpu/
  RenderState/present_wgpu
- frame/mod.rs (225) — ciclo: BEGIN/END_FRAME + SNAPSHOT + drenar + re-exports
- frame/render.rs (239) — InlineStyle/render_dom/render_block/_body/render_inline/
  collect_text

API pública intacta: lib.rs (`mod frame; pub use frame::*;`), app.rs, ctx.rs com
zero diff. Visibilidade ajustada: present_wgpu/render_dom → pub(crate); mod.rs
re-exporta Backend/GpuConfig/RenderState/WindowChrome.

Kill-gate de regressão: o present_mode da surface virou a const UI_PRESENT_MODE
(= Fifo) + teste vsync_kill_gate que falha o build se trocarem para
Immediate/Mailbox — protege o fix da janela-que-parava-sozinha (sem throttle no
loop, vsync é obrigatório).

Testes: 28/28 verdes (27 + vsync_kill_gate). Build release OK. E2E incrementador
roda (render intacto pós-split).
